+++ test/cases/phase1/logical/or-returns-operand/expected.c
@@ -2,7 +2,7 @@
 
 int main(void) {
     TSC_INIT();
-    int32_t _tsc_lhs_0 = 0;
+    double _tsc_lhs_0 = 0;
     int32_t x = (_tsc_lhs_0) ? _tsc_lhs_0 : 42;
     printf("%d\n", x);
     return 0;
