+++ test/cases/phase2/type-inference/literal-i32/expected.c
@@ -2,7 +2,7 @@
 
 int main(void) {
     TSC_INIT();
-    const int32_t x = 42;
-    printf("%d\n", x);
+    const double x = 42.0;
+    printf("%g\n", x);
     return 0;
 }
