+++ test/cases/phase4/closures/closure-return-from-fn/expected.c
@@ -1,12 +1,12 @@
 #include "runtime.h"
 
-typedef struct { int32_t n; } _closure_0_env;
+typedef struct { double n; } _closure_0_env;
 
 static int32_t _closure_0_fn(_closure_0_env *env) {
     return env->n;
 }
 
-typedef struct { int32_t m; } _closure_1_env;
+typedef struct { double m; } _closure_1_env;
 
 static int32_t _closure_1_fn(_closure_1_env *env) {
     return env->m;
@@ -14,10 +14,10 @@ static int32_t _closure_1_fn(_closure_1_env *env) {
 
 int main(void) {
     TSC_INIT();
-    int32_t n = 5;
+    double n = 5.0;
     _closure_0_env add5_env = {.n = n};
     tsc_closure add5 = {.env = &add5_env, .fn = (void*)_closure_0_fn};
-    int32_t m = 10;
+    double m = 10.0;
     _closure_1_env add10_env = {.m = m};
     tsc_closure add10 = {.env = &add10_env, .fn = (void*)_closure_1_fn};
     printf("%d\n", ((int32_t (*)(void *))add5.fn)(add5.env));
