+++ test/cases/phase4/closures/closure-scope-cleanup/expected.c
@@ -1,6 +1,6 @@
 #include "runtime.h"
 
-typedef struct { int32_t a; } _closure_0_env;
+typedef struct { double a; } _closure_0_env;
 
 static int32_t _closure_0_fn(_closure_0_env *env) {
     return env->a + 1;
@@ -8,13 +8,13 @@ static int32_t _closure_0_fn(_closure_0_env *env) {
 
 int main(void) {
     TSC_INIT();
-    int32_t result = 0;
+    double result = 0.0;
     {
-        int32_t a = 10;
+        double a = 10.0;
         _closure_0_env f_env = {.a = a};
         tsc_closure f = {.env = &f_env, .fn = (void*)_closure_0_fn};
         result = ((int32_t (*)(void *))f.fn)(f.env);
     }
-    printf("%d\n", result);
+    printf("%g\n", result);
     return 0;
 }
