+++ test/cases/phase7/promise/then/expected.c
@@ -23,8 +23,8 @@ static void main_poll(main_state *self) {
             getValue_poll(&self->_await_0);
             if (!self->_await_0._done) return;
             int32_t x = self->_await_0._result;
-            const int32_t y = x * 2;
-            printf("%d\n", y);
+            const double y = x * 2;
+            printf("%g\n", y);
             self->_done = true;
             return;
     }
